move ZosmfRenameRequest test in model project
