+++ data-sets-api-server/src/test/java/org/zowe/data/sets/services/zosmf/PutDataSetRenameZosmfRequestRunnerTest.java
@@ -22,11 +22,9 @@ import org.zowe.api.common.exceptions.HtmlEscapedZoweApiRestException;
 import org.zowe.api.common.exceptions.ZoweApiErrorException;
 import org.zowe.api.common.test.services.zosmf.AbstractZosmfRequestRunnerTest;
 import org.zowe.data.sets.model.DataSetRenameRequest;
-import org.zowe.data.sets.model.ZosmfRenameRequest;
 
 import java.io.IOException;
 
-import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.when;
 
 @RunWith(PowerMockRunner.class)
@@ -163,27 +161,4 @@ public class PutDataSetRenameZosmfRequestRunnerTest extends AbstractZosmfRequest
         return testObj;
     }
     
-    @Test
-    public void test_ZosmfRenameRequest_model() throws Exception {
-        String dsn = "TEST.JCL"; 
-        String withoutMember = dsn;
-        
-        String withoutMemberReqString = String.format("{\"request\":\"rename\",\"from-dataset\":{\"dsn\":\"%s\"}}", dsn);
-        
-        ZosmfRenameRequest zosmfRenameReq = ZosmfRenameRequest.createFromDataSetRenameRequest(withoutMember);
-        assertEquals(dsn, zosmfRenameReq.getDsn());
-        assertEquals("", zosmfRenameReq.getMember());
-        assertEquals(withoutMemberReqString, zosmfRenameReq.buildJson().toString());
-        assertEquals(withoutMemberReqString, zosmfRenameReq.toString());
-        
-        String member = "OLD";
-        String withMember = String.format("%s(%s)", dsn, member);
-        String withtMemberReqString = String.format("{\"request\":\"rename\",\"from-dataset\":{\"dsn\":\"%s\",\"member\":\"%s\"}}", dsn, member);
-        
-        zosmfRenameReq = ZosmfRenameRequest.createFromDataSetRenameRequest(withMember);
-        assertEquals(dsn, zosmfRenameReq.getDsn());
-        assertEquals(member, zosmfRenameReq.getMember());
-        assertEquals(withtMemberReqString, zosmfRenameReq.buildJson().toString());
-        assertEquals(withtMemberReqString, zosmfRenameReq.toString());
-    }
 }
